refactor: endpoint를 shared/api에 통합
