+++ frontend/src/features/mindmap/hooks/useDeleteMindmap.ts
@@ -1,8 +1,8 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 import { ApiError } from "@/features/auth/types/api";
-import { mindmapEndpoints } from "@/features/mindmap/api/mindmap_endpoints";
 import { mindmapKeys } from "@/features/mindmap/api/mindmap_query_keys";
+import { mindmapEndpoints } from "@/shared/api/api";
 import { del } from "@/shared/api/method";
 
 const fetchDeleteMindmap = (mindmapId: string) => {
