+++ frontend/src/features/mindmap/hooks/useUpdateMindmapFavorite.ts
@@ -1,8 +1,8 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 
 import { ApiError } from "@/features/auth/types/api";
-import { mindmapEndpoints } from "@/features/mindmap/api/mindmap_endpoints";
 import { mindmapKeys } from "@/features/mindmap/api/mindmap_query_keys";
+import { mindmapEndpoints } from "@/shared/api/api";
 import { patch } from "@/shared/api/method";
 
 type UpdateMindmapFavoriteParams = {
