+++ frontend/src/shared/api/api.ts
@@ -1,2 +1,13 @@
 export const USER_ENDPOINT = "/users";
 export const USER_ME_ENDPOINT = `${USER_ENDPOINT}/me`;
+
+export const mindmapEndpoints = {
+    create: "/mindmaps",
+    list: () => "/mindmaps",
+    detail: (mindmapId: string) => `/mindmaps/${mindmapId}`,
+    delete: (mindmapId: string) => `/mindmaps/${mindmapId}`,
+    rename: (mindmapId: string) => `/mindmaps/${mindmapId}/name`,
+    favorite: (mindmapId: string) => `/mindmaps/${mindmapId}/favorite`,
+
+    node: (mindmapId: string, nodeId: string) => `/mindmaps/${mindmapId}/nodes/${nodeId}`,
+} as const;
